LeaderBoard/requirements.txt
Package           Version
----------------- ------------
bcrypt            4.0.1
blinker           1.7.0
click             8.1.7
Flask             3.0.0
Flask-Bcrypt      1.0.1
Flask-Login       0.6.3
Flask-SQLAlchemy  3.1.1
Flask-WTF         1.2.1
greenlet          3.0.1
itsdangerous      2.1.2
Jinja2            3.1.2
MarkupSafe        2.1.3
numpy             1.26.2
pandas            2.1.3
Pillow            10.1.0
pip               23.3
python-dateutil   2.8.2
pytz              2023.3.post1
setuptools        68.0.0
six               1.16.0
SQLAlchemy        2.0.23
typing_extensions 4.8.0
tzdata            2023.3
Werkzeug          3.0.1
wheel             0.41.2
WTForms           3.1.1
